requirements.txt
absl-py==2.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2025.1.31
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.1.7
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-yasg==1.21.5
entrypoints==0.4
Flask==2.2.3
flatbuffers==25.1.24
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.13.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-codec==1.3.2
opt_einsum==3.4.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.6.0
pymongo==4.3.3
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.3.5
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
streamlit==1.22.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
Werkzeug==2.2.3
wrapt==1.17.2
yarl==1.9.2
zipp==3.15.0
python-multipart
requests 
torch>=1.8.1
torchvision>=0.9.1
torchaudio>=0.8.1
transformers==4.8.1
sentencepiece>=0.1.91
uvicorn==0.22.0
fastapi==0.68.0
tokenizers==0.10.3
typing-extensions>=3.10.0,<4.6.0
git+https://github.com/SKTBrain/KoBERT.git@master
openai-whisper
pydub